+++ R/transform.R
@@ -179,5 +179,5 @@ usmap_crs <- function() {
   }
 
   sp::CRS("+proj=laea +lat_0=45 +lon_0=-100 +x_0=0 +y_0=0
-                     +a=6370997 +b=6370997 +units=m +no_defs")
+          +a=6370997 +b=6370997 +units=m +no_defs")
 }