+++ revdep/checks.noindex/refuge/old/refuge.Rcheck/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(refuge)
+
+test_check("refuge")